engine_traits: Move DBOptions into DBOptionsExt
