+++ components/engine_rocks/src/lib.rs
@@ -15,9 +15,9 @@ mod iterator;
 pub use self::iterator::Iterator;
 mod cf;
 pub use cf::*;
-mod db_options;
 mod options;
 mod util;
+mod db_options;
 pub use db_options::*;
 mod cf_options;
 pub use cf_options::*;
