+++ components/engine_traits/src/lib.rs
@@ -21,8 +21,8 @@ mod engine;
 pub use crate::engine::*;
 mod options;
 pub use crate::options::*;
-mod db_options;
 mod util;
+mod db_options;
 pub use crate::db_options::*;
 mod cf_options;
 pub use crate::cf_options::*;
